Refactor the old analysis gui package name
