+++ src/main/oldWindowConstruction/testSmellPanel/ClassWithGeneralFixturePanel.java
@@ -1,4 +1,4 @@
-package main.windowConstruction.testSmellPanel;
+package main.oldWindowConstruction.testSmellPanel;
 
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
@@ -18,7 +18,7 @@ import main.refactor.IRefactor;
 import main.refactor.strategy.GeneralFixtureStrategy;
 import main.testSmellDetection.testSmellInfo.generalFixture.GeneralFixtureInfo;
 import main.testSmellDetection.testSmellInfo.generalFixture.MethodWithGeneralFixture;
-import main.windowConstruction.ContextualAnalysisFrame;
+import main.oldWindowConstruction.ContextualAnalysisFrame;
 
 import javax.swing.*;
 import javax.swing.border.Border;
